backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
openai==1.52.2
google-generativeai==0.8.3
python-docx==1.1.2
pypdf==5.0.1
alembic==1.13.3
httpx==0.27.2
orjson==3.10.7
